refactor: 🏄🏽  move point related logic to separate module
