+++ src/domain/mod.rs
@@ -1,0 +1,3 @@
+mod point;
+
+pub use point::parse_points_file;
